Subgroup Analysis
New Script SME_script, in which the subgroups are analysed. Warning,
produces lots of figures. Many comments, much wow.
For this script to work, I also include attributes_counter_mat and
simpleDescriptionGraphGenerator_mat which are equivalents of the
eponymous but without _mat functions which work on Data as a cell.
simiStrCmpRefined is the function creating the similarity matrix.